Extract language localization logic into separate utility
