+++ src/main/java/org/folio/fqm/service/ResultSetService.java
@@ -1,11 +1,8 @@
 package org.folio.fqm.service;
 
-import tools.jackson.core.json.JsonReadFeature;
-import tools.jackson.core.type.TypeReference;
 import tools.jackson.databind.JsonNode;
 import tools.jackson.databind.ObjectMapper;
 import tools.jackson.databind.node.ObjectNode;
-import tools.jackson.databind.json.JsonMapper;
 import com.jayway.jsonpath.JsonPath;
 import lombok.extern.log4j.Log4j2;
 import org.apache.commons.collections4.CollectionUtils;
@@ -24,7 +21,6 @@ import org.springframework.stereotype.Service;
 import lombok.RequiredArgsConstructor;
 
 import java.io.IOException;
-import java.io.InputStream;
 import java.sql.Timestamp;
 import java.time.Instant;
 import java.time.ZoneId;
@@ -37,7 +33,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -56,11 +51,8 @@ public class ResultSetService {
     .toFormatter().withZone(ZoneOffset.UTC); // force interpretation as UTC
   private static final String DATE_TIME_REGEX = "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}([+-]\\d{2}:\\d{2}(:\\d{2})?|Z|)$";
   private static final String COUNTRY_TRANSLATION_TEMPLATE = "mod-fqm-manager.countries.%s";
-  private static final String LANGUAGE_DISAMBIGUATION_TEMPLATE = "mod-fqm-manager.languages.disambiguated";
-  private static final String LANGUAGES_FILEPATH = "languages.json5";
   private static final String NESTED_FIELD_MARKER = "[*]->";
   private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
-  private static final LanguageMetadata LANGUAGE_METADATA = loadLanguageMetadata();
 
   private final ResultSetRepository resultSetRepository;
   private final EntityTypeFlatteningService entityTypeFlatteningService;
@@ -319,7 +311,7 @@ public class ResultSetService {
           return null;
         }
         String rawCode = value.toString();
-        return languageDisplayMap.getOrDefault(rawCode, localizeLanguageCode(rawCode, folioLocale));
+        return languageDisplayMap.getOrDefault(rawCode, LanguageLocalizationUtils.localizeLanguageCode(rawCode, folioLocale));
       })
       .toList();
   }
@@ -328,61 +320,11 @@ public class ResultSetService {
     try {
       String rawJson = languageClient.get(executionContext.getTenantId());
       List<String> codes = JsonPath.parse(rawJson).read("$.facets.languages.values.*.id");
-      List<LocalizedLanguageValue> localizedValues = codes.stream()
-        .filter(StringUtils::isNotEmpty)
-        .distinct()
-        .map(code -> new LocalizedLanguageValue(code, localizeLanguageCode(code, folioLocale)))
-        .toList();
-
-      Map<String, Long> distinctRawValueCountsByLabel = localizedValues.stream()
-        .filter(item -> item.localizedValue() != null)
-        .map(LocalizedLanguageValue::localizedValue)
-        .distinct()
-        .collect(Collectors.toMap(
-          Function.identity(),
-          label -> localizedValues.stream()
-            .filter(other -> label.equals(other.localizedValue()))
-            .map(LocalizedLanguageValue::rawValue)
-            .filter(Objects::nonNull)
-            .distinct()
-            .count()
-        ));
-
-      return localizedValues.stream()
-        .collect(Collectors.toMap(
-          LocalizedLanguageValue::rawValue,
-          item -> {
-            long distinctCount = distinctRawValueCountsByLabel.getOrDefault(item.localizedValue(), 0L);
-            if (distinctCount > 1) {
-              return translationService.format(
-                LANGUAGE_DISAMBIGUATION_TEMPLATE,
-                "label", item.localizedValue(),
-                "code", item.rawValue()
-              );
-            }
-            return item.localizedValue();
-          }
-        ));
+      return LanguageLocalizationUtils.getLanguageDisplayMap(codes, folioLocale, translationService);
     } catch (Exception e) {
       log.warn("Failed to retrieve tenant language labels for result localization. Falling back to per-value localization.", e);
       return Map.of();
     }
-  }
-
-  private String localizeLanguageCode(String code, Locale folioLocale) {
-    String a2Code = LANGUAGE_METADATA.codeToA2Map().get(code);
-    String name = LANGUAGE_METADATA.codeToNameMap().get(code);
-    if (StringUtils.isNotEmpty(a2Code)) {
-      Locale languageLocale = new Locale(a2Code);
-      String label = languageLocale.getDisplayLanguage(folioLocale);
-      if (StringUtils.isNotEmpty(label) && !"Undetermined".equals(label)) {
-        return label;
-      }
-    }
-    if (StringUtils.isNotEmpty(name) && !"Undetermined".equals(name)) {
-      return name;
-    }
-    return code;
   }
 
   private Locale getFolioLocale() {
@@ -403,37 +345,6 @@ public class ResultSetService {
       .toList();
   }
 
-  private static LanguageMetadata loadLanguageMetadata() {
-    ObjectMapper mapper = JsonMapper
-      .builder()
-      .enable(JsonReadFeature.ALLOW_SINGLE_QUOTES)
-      .enable(JsonReadFeature.ALLOW_UNQUOTED_PROPERTY_NAMES)
-      .build();
-
-    try (InputStream input = ResultSetService.class.getClassLoader().getResourceAsStream(LANGUAGES_FILEPATH)) {
-      List<Map<String, String>> languages = mapper.readValue(input, new TypeReference<>() {});
-      Map<String, String> codeToNameMap = new HashMap<>();
-      Map<String, String> codeToA2Map = new HashMap<>();
-
-      for (Map<String, String> language : languages) {
-        String alpha3 = language.get("alpha3");
-        String alpha2 = language.get("alpha2");
-        String name = language.get("name");
-
-        codeToA2Map.put(alpha3, alpha2);
-        codeToNameMap.put(alpha3, name);
-
-        if (StringUtils.isNotEmpty(alpha2)) {
-          codeToA2Map.put(alpha2, alpha2);
-          codeToNameMap.put(alpha2, name);
-        }
-      }
-      return new LanguageMetadata(codeToNameMap, codeToA2Map);
-    } catch (IOException e) {
-      throw new IllegalStateException("Failed to load language metadata", e);
-    }
-  }
-
   private static String adjustDate(Instant instant, ZoneId tenantTimezone) {
     return instant.atZone(tenantTimezone).toLocalDate().toString();
   }
@@ -446,8 +357,4 @@ public class ResultSetService {
     log.warn("Database date value is in an unrecognized format: \"{}\"", value);
     return value;
   }
-
-  private record LanguageMetadata(Map<String, String> codeToNameMap, Map<String, String> codeToA2Map) {}
-
-  private record LocalizedLanguageValue(String rawValue, String localizedValue) {}
 }
